Change Call Detail id field to readonly
The client was being allowed to enter the id field of a Call Detail Record.
This made the field mandatory in the payload of create post.
Now the id field is not mandatory in the creation payload.
If it is informed it will be ignored.
The id will always be determined automatically by the API.